generate_taxon_counts and compute_merged_taxon_counts moved to blast.rs
